// src/app/Models/orderModel.ts
export class orderModel {
  id: number;
  orderDate: Date;
  deliveryDate: Date;
  pickupDate:Date;
  Township_id:number;
  clientName: string;
  clitentPhone: string;
  clitentFlatNo: string;
  clientAddress: string;
  deliveryCharegs: number;
  riderID: number;
  riderEarning: number;
  operatorID: number;
  status:string;
  riderTakeOption:string;
  longitude:string;
  latitude:String;
}